pyproject.toml
[project]
name = "electrakan"
version = "0.1.0"
description = "ElectraKAN is Electra model using KAN model instead of Fully Connected Layer"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    {name = "shin jung tae", email = "shinjeongtae@gmail.com"}
]
dependencies = [
    "tqdm>=4.67.1",
    "torch>=2.5.1",
    "lightning>=2.3.1",
    "hydra-core>=1.3.2",
    "ruamel.yaml>=0.18.3",
    "omegaconf>=2.3.0",
    "torchmetrics>=1.1.1",
    "transformers>=4.45.2",
    "ujson>=5.7.0",
    "polars>=1.13.0",
    "datasets>=3.3.1",
    "orjson>=3.10.0",
    "pyarrow>=19.0.1",
]

[dependency-groups]
dev = [
    "ipywidgets",
    "pytest",
    "neuronx-cc",
    "torch-neuronx",
    "marimo",
    "polars",
    "line-profiler",
    "memory-profiler",
    "ipykernel",
    "ipython",
    "jupyter",
    "jupyterlab",
    "jupyterlab-git"
]

[build-system]
requires = ["setuptools", "wheel", "poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "electrakan"
version = "0.1.0"
description = "ElectraKAN is Electra model using KAN model instead of Fully Connected Layer"
authors = ["shin jung tae <shinjeongtae@gmail.com>"]
packages = [
    { include = "ElectraKAN", from = "src" }
]

[tool.poetry.scripts]
train = "ElectraKAN.train:main"